Re-org of the 'service' package to delineate classes
