+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/LegacyAccountDetailsResponseDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.extern.jackson.Jacksonized;
 import uk.gov.hmcts.opal.dto.AccountDetailsDto;
 import uk.gov.hmcts.opal.dto.ToJsonString;
-import uk.gov.hmcts.opal.service.DefendantAccountService;
+import uk.gov.hmcts.opal.service.opal.DefendantAccountService;
 
 import java.util.Comparator;
 import java.util.List;
