+++ src/main/java/uk/gov/hmcts/opal/service/legacy/LegacyGatewayResponseException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.service;
+package uk.gov.hmcts.opal.service.legacy;
 
 public class LegacyGatewayResponseException extends RuntimeException {
