+++ src/main/java/uk/gov/hmcts/opal/service/opal/PartyService.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.service;
+package uk.gov.hmcts.opal.service.opal;
 
 
 import lombok.RequiredArgsConstructor;
@@ -7,6 +7,7 @@ import uk.gov.hmcts.opal.dto.AccountSearchDto;
 import uk.gov.hmcts.opal.dto.PartyDto;
 import uk.gov.hmcts.opal.entity.PartySummary;
 import uk.gov.hmcts.opal.repository.PartyRepository;
+import uk.gov.hmcts.opal.service.PartyServiceInterface;
 
 import java.util.List;
