+++ src/main/java/uk/gov/hmcts/opal/service/proxy/NoteServiceProxy.java
@@ -1,14 +1,18 @@
-package uk.gov.hmcts.opal.service;
+package uk.gov.hmcts.opal.service.proxy;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.opal.dto.NoteDto;
+import uk.gov.hmcts.opal.service.DynamicConfigService;
+import uk.gov.hmcts.opal.service.NoteServiceInterface;
+import uk.gov.hmcts.opal.service.legacy.LegacyNoteService;
+import uk.gov.hmcts.opal.service.opal.NoteService;
 
 @Service
 @RequiredArgsConstructor
 @Qualifier("noteServiceProxy")
-public class NoteServiceProxy implements NoteServiceInterface, LegacyProxy {
+public class NoteServiceProxy implements NoteServiceInterface, ProxyInterface {
 
     private final NoteService opalNoteService;
     private final LegacyNoteService legacyNoteService;
