+++ src/main/java/uk/gov/hmcts/opal/service/proxy/ProxyInterface.java
@@ -1,7 +1,9 @@
-package uk.gov.hmcts.opal.service;
+package uk.gov.hmcts.opal.service.proxy;
+
+import uk.gov.hmcts.opal.service.DynamicConfigService;
 
 /** A marker interface to help identify code that is to be removed at a future date. */
-public interface LegacyProxy {
+public interface ProxyInterface {
 
     public static final String LEGACY_MODE = "LEGACY";
