+++ src/test/java/uk/gov/hmcts/opal/controllers/NoteControllerTest.java
@@ -8,7 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.opal.dto.NoteDto;
-import uk.gov.hmcts.opal.service.NoteService;
+import uk.gov.hmcts.opal.service.opal.NoteService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
