+++ src/test/java/uk/gov/hmcts/opal/controllers/PartyControllerTest.java
@@ -8,7 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.opal.dto.PartyDto;
-import uk.gov.hmcts.opal.service.PartyService;
+import uk.gov.hmcts.opal.service.opal.PartyService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
